add tests for all valid date input formats
The big change in this commit is the new
`test/ex_ical_date_format_test.ex` test. It tests all 6 valid date
formats against both a TZID-free calendar and a calendar with the global
TZID set.

The testcases are declared in a comprehension because this makes
adding/removing/managing input formats and desired outputs trivial
(versus managing a dozen testcases).