remove invalid type
`Timex.Typex` was introduced in v2.0, but we're locked onto `"~> 1.0"`.
Thus, dialyzer couldn't find `Timex.Types.valid_timezone`.

For now, I just added a local implementation of `valid_timezone`; later
on, when `timex` is upgraded to `>= 2.0`, that can be switched back to
`Timex.Types.valid_timezone` (to keep in sync with the valid timezone
formats that timex accepts).