Flesh out documentation (#11)
* update credo

Credo was showing a warning message, so I upgraded. This shouldn't
affect behavior or warnings.

* make `ExIcal.Utils` functions private

These two functions: `date_before?/2` and `date_after?/2`, are only used
internally in the `ExIcal.Utils` module. They aren't delegated to the
`ExIcal` module, and their API (`ExIcal.Utils.date_before/2`, etc)
suggests that we don't expect developers to use that API.

Making these two functions private until a compelling reason is given to
publish them means that we can change the internal API with a breaking
change, and it's less public interface to maintain. It'll be easier to
make these public later on than to hide them.

* use `alias` instead of `import`

`import` imports all of a module's functions into the local namespace.
Using `alias` instead gives finer control over what exists in the
namespace.

Additionally, this change makes it clearer how we're using `Timex`. When
we go to refactor that out (see #8), this should make it a bit easier.

* add typespecs

The API is stable-ish, so I added typespecs. This means that the `s`
command in iex will list the public methods, and the generated exdocs
will be more detailed.

Many of these specs do use Timex-derived structs (`%Timex.DateTime{}`),
so in the course of #9, these will need to be changed.

* consolidate aliases

The code was a bit inconsistent in which aliases were declared, and some
submodules under the same module were declared on multiple lines (as
opposed to the `ExIcal.{SubOne,SubTwo}` syntax).

Now all the `Timex` and `ExIcal` repeated use aliases are declared at
the top of the file, and they all use the curly bracket notation for
consistency. I hope that this makes it easy to see the major file
dependencies in one place as well as makes the code more concise.

* flesh out documentation

This commit adds/fleshes out documentation in all modules and public
interface functions. My goal here was to make the library look more
"polished" and to help users find answers to their questions related to
`ExIcal` or the iCal spec. I tested it out in `iex` as well as the
generated docs, and everything looks as intended.

Changes:

  - add/improve lots of documentation
  - tidy code in `lib/ex_ical/utils` (break up long line)

Decisions & Justifications:

  - **lots of links**: in `lib/ex_ical/event.ex`, I included a lot of links
    to the iCal spec. They render as expected in the generated docs, but
    they show up as an ugly-ish list in `iex`. I think it's still worth
    it to have them since they'll help people with questions about the
    iCal property rules.

  - **ask for input**: in `lib/ex_ical/event.ex` and
    `lib/ex_ical/parser.ex`, I included paragraphs mentioning that the
    spec was not fully implemented and to file an issue if they need
    some unsupported property. I think it's valuable to have this in the
    docs because that's where people will go if they're looking for
    property support, but it might be too conversational for your taste.